refactor inclusion report
