+++ ui/apps/vue-mri-ui-lib/src/query-filter/components/InclusionReport/components/FunnelChart.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+defineProps<{
+  funnelChartRef: HTMLElement | null
+}>()
+</script>
+
+<template>
+  <div class="chart-section">
+    <h4>Attrition visualization</h4>
+    <div :ref="(el) => funnelChartRef = el as HTMLElement" class="funnel-chart"></div>
+  </div>
+</template>
+
