+++ ui/apps/vue-mri-ui-lib/src/query-filter/components/InclusionReport/components/TreemapChart.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+defineProps<{
+  treemapChartRef: HTMLElement | null
+}>()
+</script>
+
+<template>
+  <div class="chart-section">
+    <h4>Population visualization</h4>
+    <div :ref="(el) => treemapChartRef = el as HTMLElement" class="treemap-chart"></div>
+  </div>
+</template>
+
