+++ ui/apps/vue-mri-ui-lib/src/query-filter/styles/InclusionReport.scss
@@ -160,7 +160,6 @@ h4 {
     color: #333;
   }
 }
-
 .chart-section {
   display: flex;
   flex-direction: column;
